requirements.txt
Django~=5.0.2amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.3
Django==5.0.2
django-celery-beat==2.5.0
django-redis==5.4.0
django-rest-framework==0.1.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-simplejwt-sso==1.0.1
kombu==5.3.5
prompt-toolkit==3.0.43
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13

pyotp~=2.9.0